Axed AppDelegate, refactored layouts, fixed warnings. Style audit. Purged crusty/superfluous files.
• Eliminated unnecessary/distracting AppDelegate
• Cleaned up Nibs/ layouts.
• Got rid of TONS of redundant or incorrect property specifications
(assign, nonatomic, blah blah blah).
• Eliminated separate view files, by instead strategically placing the
views nearer to their controllers - to better show the balance between
private and public declaration, property visibility, etc.
• Also made some rather opinionated changes to the formatting (mostly)
- of the RAC constructs.

> I've been watching this project for a while... and have always wanted
to help... cause well, the idea is great, but I'd imagine that the way
some things are - I dunno, "spaced".. is like some LSD trip.  Even if
you get it...  you can't read it, lol.  This is my judicious and
methodical attempt at presenting another possible way to "do it".  One
which I hope is easier to read, easier to comprehend, and for us
dumb-dumbs... more approachable.

> _I know structural suggestions, (which this commit isn't / nor is
intended to be) are often disregarded as stylistic choices.. but as a
consumer of the API... I edited this project to stay exactly the same,
but to just be simpler and clearer, and easier to get one's head
around.  A lot of what I did was get rid of unnecessary stuff, like all
those synthesize statements, unnecessary brackets, etc. They make the
good stuff - the meat and potatoes, etc harder to find, understand, and
use.  Ok, let me know if you have any questions or comments, or
F-You's.  :-)_